requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
faster-whisper==1.1.1
numpy==2.2.4
websockets==16.0
python-dotenv==1.0.1
pystray==0.19.5
Pillow==11.1.0
